# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Portfolio API",
        "description": "This is the Swagger UI documentation for my Laravel API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/social": {
            "get": {
                "tags": [
                    "Social"
                ],
                "summary": "Get all social",
                "operationId": "b41415bd1d9fac5aebdc656456f948dd",
                "responses": {
                    "200": {
                        "description": "List of social"
                    }
                }
            },
            "post": {
                "tags": [
                    "Social"
                ],
                "summary": "Create a new social",
                "operationId": "ec65200a160f84e763419e7f07d91883",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "s_img",
                                    "s_title",
                                    "s_link",
                                    "display"
                                ],
                                "properties": {
                                    "s_img": {
                                        "type": "integer"
                                    },
                                    "s_title": {
                                        "type": "string"
                                    },
                                    "s_link": {
                                        "type": "string"
                                    },
                                    "display": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Social created successfully"
                    }
                }
            }
        },
        "/api/social/{id}": {
            "get": {
                "tags": [
                    "Social"
                ],
                "summary": "Get an social by ID",
                "operationId": "1a08ef667a0e8908c634027e5af00b92",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Social retrieved successfully"
                    },
                    "404": {
                        "description": "Social not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Social"
                ],
                "summary": "Delete an social by ID",
                "operationId": "8645405a1b0a4f008a0180e9a7209e16",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "delete social successfully"
                    },
                    "404": {
                        "description": "Social not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Social"
                ],
                "summary": "Update a social by ID",
                "operationId": "08fbea25d093404e660df4fd7fd0361d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "s_img": {
                                        "type": "integer"
                                    },
                                    "s_title": {
                                        "type": "string"
                                    },
                                    "s_link": {
                                        "type": "string"
                                    },
                                    "display": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Social updated successfully"
                    },
                    "404": {
                        "description": "Social not found"
                    }
                }
            }
        },
        "/api/text": {
            "get": {
                "tags": [
                    "Text"
                ],
                "summary": "Get all text",
                "operationId": "510660bc69a2531e75afe731e7bc7579",
                "responses": {
                    "200": {
                        "description": "List of text"
                    }
                }
            },
            "post": {
                "tags": [
                    "Text"
                ],
                "summary": "Create a new text",
                "operationId": "3ef275b33da99a94cf7d66628d8ce8c3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "t_detail",
                                    "t_type",
                                    "display"
                                ],
                                "properties": {
                                    "t_detail": {
                                        "type": "string"
                                    },
                                    "t_type": {
                                        "type": "integer"
                                    },
                                    "display": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Text created successfully"
                    }
                }
            }
        },
        "/api/text/{id}": {
            "get": {
                "tags": [
                    "Text"
                ],
                "summary": "Get an text by ID",
                "operationId": "298b61dd3f876b4448b506d013499948",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Text retrieved successfully"
                    },
                    "404": {
                        "description": "Text not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Text"
                ],
                "summary": "Delete an text by ID",
                "operationId": "3d1e80bd2127fbe72e6b9159f89f9b80",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "delete text successfully"
                    },
                    "404": {
                        "description": "Text not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Text"
                ],
                "summary": "Update a text by ID",
                "operationId": "e82e01a1e087e3040470e9b013cb461c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "t_detail": {
                                        "type": "string"
                                    },
                                    "t_type": {
                                        "type": "integer"
                                    },
                                    "display": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Text updated successfully"
                    },
                    "404": {
                        "description": "Text not found"
                    }
                }
            }
        }
    },
...
}